requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
pymongo>=4.5.0
motor>=3.3.2
python-dotenv>=1.0.0
pydantic>=2.4.2
openai>=1.12.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.1
numpy>=1.26.0
pillow>=10.0.1
pypdf>=4.0.0
python-docx>=1.0.0
uuid>=1.30.0
pytesseract>=0.3.10
bson>=0.5.10